+++ demisto_sdk/commands/common/tools.py
@@ -161,12 +161,13 @@ def run_command(command, is_silenced=True, exit_on_error=True, cwd=None):
     return output
 
 
-def get_remote_file(full_file_path, tag='master', return_content=False):
+def get_remote_file(full_file_path, tag='master', return_content=False, suppress_print=False):
     """
     Args:
         full_file_path (string):The full path of the file.
         tag (string): The branch name. default is 'master'
         return_content (bool): Determines whether to return the file's raw content or the dict representation of it.
+        suppress_print (bool): whether to suppress the warning message in case the file was not found.
     Returns:
         The file content in the required format.
 
@@ -180,9 +181,10 @@ def get_remote_file(full_file_path, tag='master', return_content=False):
         res = requests.get(github_path, verify=False, timeout=10)
         res.raise_for_status()
     except Exception as exc:
-        print_warning('Could not find the old entity file under "{}".\n'
-                      'please make sure that you did not break backward compatibility. '
-                      'Reason: {}'.format(github_path, exc))
+        if not suppress_print:
+            print_warning('Could not find the old entity file under "{}".\n'
+                          'please make sure that you did not break backward compatibility. '
+                          'Reason: {}'.format(github_path, exc))
         return {}
     if return_content:
         return res.content
@@ -194,6 +196,23 @@ def get_remote_file(full_file_path, tag='master', return_content=False):
     else:
         details = {}
     return details
+
+
+def filter_files_on_pack(pack: str, file_paths_list=str()) -> set:
+    """
+    filter_files_changes_on_pack.
+
+    :param file_paths_list: list of content files
+    :param pack: pack to filter
+
+    :return: files_paths_on_pack: set of file paths contains only files located in the given pack
+    """
+    files_paths_on_pack = set()
+    for file in file_paths_list:
+        if get_pack_name(file) == pack:
+            files_paths_on_pack.add(file)
+
+    return files_paths_on_pack
 
 
 def filter_packagify_changes(modified_files, added_files, removed_files, tag='master'):
@@ -653,6 +672,29 @@ def get_pack_names_from_files(file_paths, skip_file_types=None):
                 packs.add(pack)
 
     return packs
+
+
+def filter_files_by_type(file_paths=None, skip_file_types=None) -> set:
+    """get set of files and return the set whiteout the types to skip
+
+    Args:
+    - file_paths (set): set of content files.
+    - skip_file_types List[str]: list of file types to skip.
+
+    Returns:
+    files (set): list of files whiteout the types to skip
+    """
+    if file_paths is None:
+        file_paths = set()
+    files = set()
+    for path in file_paths:
+        # renamed files are in a tuples - the second element is the new file name
+        if isinstance(path, tuple):
+            path = path[1]
+        file_type = find_type(path)
+        if file_type not in skip_file_types and is_file_path_in_pack(path):
+            files.add(path)
+    return files
 
 
 def pack_name_to_path(pack_name):
@@ -1122,19 +1164,6 @@ def retrieve_file_ending(file_path: str) -> str:
         if file_ending and '.' in file_ending:
             return file_ending[1:]
     return ''
-
-
-def get_depth(data: Any) -> int:
-    """
-    Returns the depth of a data object
-    :param data: The data
-    :return: The depth of the data object
-    """
-    if data and isinstance(data, dict):
-        return 1 + max(get_depth(data[key]) for key in data)
-    if data and isinstance(data, list):
-        return 1 + max(get_depth(element) for element in data)
-    return 0
 
 
 def is_test_config_match(test_config: dict, test_playbook_id: str = '', integration_id: str = '') -> bool: