+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
+* Added validation for approved content pack usecases and tags.
+* Added new code validations for *CommonServerPython* import to `XSOAR-linter`.
+* Added *default value* and *predefined values* to argument description in **generate-docs** command.
+* Added a new validation that checks if *get-mapping-fields* command exists if the intgeration schema has *{ismappable: true}* in **validate** command.
+* Fixed an issue where the *--staged* flag recognised added files as modified in the **validate** command.
+* Fixed an issue where a backwards compatibility warning was raised for all added files in the **validate** command.
+* Fixed an issue where **validate** command failed when no tests were given for a partner supported pack.
+* Updated the **download** command to support mappers.
+
+# v1.2.9
 * Fixed an issue in the **openapi_codegen** command where it created duplicate functions name from the swagger file.
 * Fixed an issue in the **update-release-notes** command where the *update type* argument was not verified.
 * Fixed an issue in the **validate** command where no error was raised in case a non-existing docker image was presented.
@@ -7,6 +17,8 @@
 * Fixed an issue in the **lint** command where *finally* clause was not supported in main function.
 * Fixed an issue in the **validate** command where changing any entity ID was not validated.
 * Fixed an issue in the **validate** command where *--staged* flag did not bring only changed files.
+* Fixed the **update-release-notes** command to ignore changes in the metadata file.
+* Fixed the **validate** command to ignore metadata changes when checking if a version bump is needed.
 
 
 # v1.2.8